requirements.txt
absl-py==0.15.0
ai2thor==2.5.0
aws-requests-auth==0.4.3
backcall==0.2.0
botocore==1.21.60
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colour==0.1.5
cycler==0.10.0
decorator==5.1.0
Flask==2.0.2
greenlet==1.1.2
idna==3.3
imageio==2.9.0
ipython==7.28.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
kiwisolver==1.3.2
libtmux==0.10.1
machine_common_sense==0.4.6
MarkupSafe==2.0.1
marshmallow==3.13.0
matplotlib==3.4.3
matplotlib-inline==0.1.3
msgpack==1.0.2
natsort==7.1.1
neovim==0.3.1
networkx==2.6.3
numpy==1.21.2
opencv-python==4.5.3.56
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
progressbar2==3.54.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
pynvim==0.4.3
pyparsing==2.4.7
python-dateutil==2.8.2
python-utils==2.5.6
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.18.3
scipy==1.7.1
Shapely==1.7.1
six==1.16.0
tifffile==2021.10.12
torch==1.10.0
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
